# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API for QuickCertS.",
        "title": "QuickCertS API",
        "contact": {
            "name": "MMQ",
            "email": "mail@mmq.dev"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:33333",
    "basePath": "/api/v1",
    "paths": {
        "/apply/cert": {
            "post": {
                "description": "Provide the client with a certificate(unique key and signature) for app.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Apply"
                ],
                "summary": "Provide the client with a certificate(unique key and signature) for app.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorized token for client access. This value is set in path_to_qcs/configs/server.toml.",
                        "name": "X-Access-Token",
                        "in": "header"
                    },
                    {
                        "description": "Apply certificate information",
                        "name": "applyInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ApplyCertInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ApplyCertResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/apply/temp-permit": {
            "post": {
                "description": "Allow users to apply for temporary use permits on devices.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Apply"
                ],
                "summary": "Allow users to apply for temporary use permits on devices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorized token for client access. This value is set in path_to_qcs/configs/server.toml.",
                        "name": "X-Access-Token",
                        "in": "header"
                    },
                    {
                        "description": "Apply temporary permit information",
                        "name": "applyInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ApplyTempPermitInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ApplyTempPermitResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.ApplyCertInfo": {
            "type": "object",
            "required": [
                "board_name",
                "board_producer",
                "mac_address",
                "serial_number"
            ],
            "properties": {
                "board_name": {
                    "type": "string",
                    "example": "ROG CROSSHAIR X670E HERO"
                },
                "board_producer": {
                    "type": "string",
                    "example": "ASUSTEK COMPUTER INCORPORATION"
                },
                "mac_address": {
                    "type": "string",
                    "example": "B42499FE0000"
                },
                "serial_number": {
                    "type": "string",
                    "example": "779f-4e90-aebd-4295-881a-f8d7"
                }
            }
        },
        "model.ApplyCertResponse": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string",
                    "example": "3266cd6a16ca77f9c0f0ff9934eb0e29c4b6bb0729cde98811f9f0caf76d603c"
                },
                "signature": {
                    "type": "string",
                    "example": "MNj/g7W+X5PmirfgWl5jveV54t50+LZAPmByh5Py880pB2z67Ser0YvZ2G/mTNV4XcIrKmLy1ICFmQ1esjydhvBj1FOuTm3eTIixUIsFLxwlW2co/R6kCIjNRydB3N7L/kWv+ZwSjsSsdHqmMUleXV3OJruxeoXV8TLRCSGE4tHGEwhPULuBLn2aldIehDTgteJx1O1YNJGIcDM3NWVDjJnUA0Bjhq3oRvXWN4M23SnZZG2vT94wJIK0X5q6oNqFTupFjDVBCFcHeWoxQ5xZdPhfXF8rC/VTb4vkZZm5RIiIK1UC9XVaAsXVPEzlxVfYJ0gh+wULx8syE2QyB5GfyQ=="
                }
            }
        },
        "model.ApplyTempPermitInfo": {
            "type": "object",
            "required": [
                "board_name",
                "board_producer",
                "mac_address"
            ],
            "properties": {
                "board_name": {
                    "type": "string",
                    "example": "ROG CROSSHAIR X670E HERO"
                },
                "board_producer": {
                    "type": "string",
                    "example": "ASUSTEK COMPUTER INCORPORATION"
                },
                "mac_address": {
                    "type": "string",
                    "example": "B42499FE0000"
                }
            }
        },
        "model.ApplyTempPermitResponse": {
            "type": "object",
            "properties": {
                "remainingTime": {
                    "type": "integer",
                    "example": 604800
                },
                "status": {
                    "type": "string",
                    "example": "activated"
                }
            }
        },
...
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error message."
                }
            }
        },
...
    }
}